"fixed" backend server

+++ src/PowerNode.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 public class PowerNode {
 
 
-    private Long id;
+    private int id;
 
     private double demand;       // positive means load
     private double generation;   // positive means produced power
@@ -14,17 +14,17 @@ public class PowerNode {
 
     private boolean slack;       // reference node
     private String name;
-    private static Long nextId =  0L;
+    private static int nextId =  0;
 
     public PowerNode ()
     {
         this.id = nextId;
         nextId++;
     }
-    public PowerNode(Long id) {
+    public PowerNode(int id) {
         this.id = id;
     }
-    public PowerNode(Long id, String name)
+    public PowerNode(int id, String name)
     {
       this(id);
       this.name = name;
@@ -35,7 +35,7 @@ public class PowerNode {
     }
 
     // Getters & Setters
-    public Long getId() { return id; }
+    public int getId() { return id; }
 
     public double getDemand() { return demand; }
     public void setDemand(double demand) { this.demand = demand; }
